@@ -2,8 +2,10 @@
   "name": "lib-to-configure",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^8.2.14",
-    "@angular/common/http": "^8.2.14",
-    "@angular/core": "^8.2.14"
+    "@angular/common": "^11.1.2",
+    "@angular/core": "^11.1.2"
+  },
+  "dependencies": {
+    "tslib": "^2.0.0"
   }
 }
