@@ -2,7 +2,7 @@
  * Public API Surface of lib-to-configure
  */
 
-export * from './lib/lib-configuration';
+export * from './lib/config';
 export * from './lib/lib-to-configure.component';
 export * from './lib/lib-to-configure.module';
 export * from './lib/lib-to-configure.service';
